quiz-ui 쪽 싹다 리팩터링 했음
** 2시간에 걸친 여정...
